requirements.txt
tabula-py==2.2.0
tqdm==4.62.0
astroid==2.6.2
autopep8==1.5.4
certifi==2021.5.30
et-xmlfile==1.1.0
isort==5.9.2
jdcal==1.4.1
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
numpy==1.21.1
openpyxl @ file:///tmp/build/80754af9/openpyxl_1615411699337/work
pandas==1.3.1
pycodestyle==2.6.0
pylint==2.9.3
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
toml==0.10.2
wrapt==1.12.1
